Blog dir for $_GET requests;
I think it will be easier to have a blog/ dir in the resources
file where we can just show that blog based off of the tmeplate
and request the user has made. This blog post text will site here.
And the filename should point to some information in a database,
such as likes, views, date created, date last viewed, author, etc.